backend/requirements.txt
fastapi[standard]
pydantic
python-dotenv
psycopg2-binary
strawberry-graphql